Split publish specs per language
The publish_spec.rb file is growing too long to understand.
I'm splitting it per language to test the language specific things,
and testing generic things that are not language specific in the
publish_spec.rb file.
